seperate Provider component to another file and solve request changes Relates #90
