+++ src/context/index.js
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
 
-const UserContext = createContext();
+const FormContext = createContext();
 
-export default UserContext;
+export default FormContext;
